[FIX] pms_api_rest: allow decimal agency commission
The `commission` field of PmsPartnerInfo was an Integer datamodel,
truncating fractional commissions (e.g. an OTA charging 15.5%) coming
from or going to the REST API. Align it with the underlying Float field
on res.partner so decimals are preserved.

+++ pms_api_rest/datamodels/pms_partner.py
@@ -60,7 +60,7 @@ class PmsPartnerInfo(Datamodel):
     pricelistId = fields.Integer(required=False, allow_none=True)
     salesReference = fields.String(required=False, allow_none=True)
     saleChannelId = fields.Integer(required=False, allow_none=True)
-    commission = fields.Integer(required=False, allow_none=True)
+    commission = fields.Float(required=False, allow_none=True)
     invoicingPolicy = fields.String(required=False, allow_none=True)
     daysAutoInvoice = fields.Integer(required=False, allow_none=True)
     invoicingMonthDay = fields.Integer(required=False, allow_none=True)
